api/requirements.txt
appnope==0.1.0
backcall==0.2.0
certifi==2020.6.20
decorator==4.4.2
future==0.18.2
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
jupyter-client==6.1.3
jupyter-core==4.6.3
numpy==1.19.0
opencv-python==4.2.0.34
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pyzmq==19.0.1
six==1.15.0
torch==1.5.0
torchvision==0.6.0
traitlets==4.3.3
wcwidth==0.2.5
